requirements.txt
anthropic
python-dotenv
flask
PyPDF2
streamlit
httpx